ml-chatbot-main/requirements.txt
fastapi
PyPDF2
langchain
google-generativeai
langchain_google_genai
faiss-cpu
python-dotenv
